sync

+++ phenome.py
@@ -28,7 +28,7 @@ class Phenome(object):
         #weights_dim = connections['weight'].shape[0]
         weights = [i for i in connections['weight']]
 
-        weight = tf.constant(weights)
+        weight = tf.constant(weights,dtype=tf.float32)
 
         #weight = tf.Variable(tf.random_uniform([(input_count),(output_count)], -1, 1), name="Weight")
         bias = tf.Variable(tf.zeros([(output_count)]), name="Bias")
